pyproject.toml
[project]
name = "auchan"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.2",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
]
